pyproject.toml
[project]
name = "compare-agent"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "streamlit>=1.35.0",
]
